requirements.txt
Django>=5.0,<6.0
 djangorestframework>=3.15,<4.0
 djangorestframework-simplejwt>=5.3.1,<6.0
 python-dotenv>=1.0,<2.0